Place OPRFState and friends together with the update update loop.
Keep the state structs and their maintenance code together to
avoid cross-importing between `handler` and the old `update`
module. Export just the wrapped version through the top-level
to shorten references there and in the handlers.
