+++ src/state.rs
@@ -1,14 +1,46 @@
 //! STAR Randomness web service
-//! Epoch and key rotation
+//! Epoch and key state and its management
 
+use std::sync::{Arc, RwLock};
 use time::format_description::well_known::Rfc3339;
 use tracing::{info, instrument};
 
-use crate::handler::OPRFServer;
-use crate::handler::OPRFState;
+use ppoprf::ppoprf;
 use crate::Config;
 
+/// Internal state of the OPRF service
+pub struct OPRFServer {
+    /// oprf implementation
+    pub server: ppoprf::Server,
+    /// currently-valid randomness epoch
+    pub epoch: u8,
+    /// RFC 3339 timestamp of the next epoch rotation
+    pub next_epoch_time: Option<String>,
+}
+
+/// Shareable wrapper around the server state
+pub type OPRFState = Arc<RwLock<OPRFServer>>;
+
+impl OPRFServer {
+    /// Initialize a new OPRFServer state with the given configuration
+    pub fn new(config: &Config) -> Result<Self, ppoprf::PPRFError> {
+        // ppoprf wants a vector, so generate one from our range.
+        let epochs: Vec<u8> =
+            (config.first_epoch..=config.last_epoch).collect();
+        let epoch = epochs[0];
+        let server = ppoprf::Server::new(epochs)?;
+        Ok(OPRFServer {
+            server,
+            epoch,
+            next_epoch_time: None,
+        })
+    }
+}
+
 /// Advance to the next epoch on a timer
+/// This can be invoked as a background task to handle
+/// epoch advance and key rotation according to the
+/// given Config.
 #[instrument(skip_all)]
 pub async fn epoch_loop(state: OPRFState, config: &Config) {
     let interval =
